making updateWellControls in StandardWellsDense an iterative process
Although the part we really need is that we should update the well state
based on their targets before checking whether any constraint is broken.

The current way in the master branch, will result in something like with
1500 ORAT control, while when we check the constraints, we have the ORAT
value from previous step 100, which will cause confusion of the well
control determination process. Sometimes, it can still result in the
correct results, while for more complicated situation, it can result in
undesired final results.