when a well BHP limit is defaulted and has THP constraint
We should assume this well can run under BHP limit, while it violates
the THP constraint.

The current problem in handling this case, is that, when applying dp to
defaulted BHP limit, we get a negative BHP value to find corresponding
THP value with VFPPROD table, which is not very desirable.