saving for first successful predition running at 20181024
including several major improvement for the solution strategy

1. not re-open cross-flow for special cases through function
crossFlowAllowed(), which should be removed and the logic that re-open
the crossflow if all the perforation is cross-flowing not making sense.

2. if there is no well control event for certain wells, we just copy the
well state from the previous well state for this well, including the
control mode.

3. if the drawdown for all the perforation is in the wrong direction, we
close this well for the iteration.

4. if a well is under THP control for the first time, we will provide an
estimation of the well state based on IPR curve for it to begin with.

5. including a relaxation factor for injector during Newton iteration to
avoid to get negative rates when Newton update happens.

6. if the well can not work under BHP or THP limit, we do not switch to
that limit during the well control swtiching

7. If the well is specified to run under THP limit, while it can not
operate under THP limit, we switch to BHP limit for this case.

8. For the well under defaulted BHP limit (1.atm. for producers), and it
has a THP limit, we do not do the normal limit calculation. We will
assume that the well can work under this BHP limit while it will violate
the THP limit at the same time. The extrapolation makes the normal
calculation process fail.

There are a lot of outputting for debugging code there, which remain to
be cleaned out.