changing the logic that computes well potential for each well
All the wells should have a BHP limit, so the well potential under BHP
limit will always be calculated.

If there is a THP limit, the well potential under THP limit should also
be calculated. Then we pick the smaller value of the well potentials as
a result of the more strict limit between BHP and THP.